+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Docker to the project.
+### Added
 - Database parameters will be saved to env file instead of database config class.
 
 ## [3.3.1] - 22-08-07
@@ -14,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - ``composer env-production`` to build ``.env`` file from ``.env.production``
   - ``composer env-development`` to build ``.env`` file from ``.env.development``
 ### Changed
-- Environment defination moved from ``/public/index.php`` to ``/.env`` file.
+- Environment definition moved from ``/public/index.php`` to ``/.env`` file.
 ### Fixed
 - [NodCMS-Bundle](https://github.com/khodakhah/nodcms-bundle) issue [#87](https://github.com/khodakhah/nodcms/issues/87)