+++ CHANGELOG.md
@@ -5,12 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [3.4.0] 20-09-13
 ### Added
 - Database parameters will be saved to env file instead of database config class.
 - Activate CodeIgniter CLI 'spark'
 - New console commands to save database connection build tables
 
-## [3.3.1] - 22-08-07
+## [3.3.1] - 2022-08-07
 ### Added
 - Composer scripts
   - ``composer env-production`` to build ``.env`` file from ``.env.production``